+++ packages/react/jest.config.js
@@ -6,7 +6,9 @@ module.exports = {
     '!src/components/SuiteHeader/util/suiteHeaderData.js',
     '!src/components/FileUploader/stories/*.jsx',
     '!src/components/Table/AsyncTable/*.js?(x)',
-    '!src/components/Page/EditPage.jsx',
+    '!src/components/WizardInline/**/*.js?(x)',
+    '!src/components/Page/(EditPage|PageHero).jsx',
+    '!src/components/Dashboard/(Dashboard|CardRenderer).jsx',
   ],
   coveragePathIgnorePatterns: ['/node_modules/', '/lib/', '/coverage/'],
   coverageReporters: ['html', 'text-summary', 'lcov'],
@@ -17,23 +19,11 @@ module.exports = {
       functions: 80,
       lines: 80,
     },
-    './src/components/**/!(WizardInline|WizardHeader|FilterHeaderRow|RowActionsCell|RowActionsError|TableHead|CatalogContent|Dashboard|CardRenderer|PageHero|TableHead|ColumnResize|ColorDropdown|TimeSeriesCard|BarChartCard|DashboardGrid|DashboardEditor).jsx': {
+    './src/components/**/!(TimeSeriesCard|BarChartCard|DashboardEditor).jsx': {
       statements: 80,
       branches: 80,
       functions: 80,
       lines: 80,
-    },
-    // TODO: remove in next release
-    './src/components/WizardInline/WizardInline.jsx': {
-      branches: 70,
-    },
-    './src/components/WizardInline//WizardHeader/WizardHeader.jsx': {
-      branches: 75,
-    },
-    './src/components/Dashboard/DashboardGrid.jsx': {
-      statements: 72,
-      branches: 60,
-      lines: 71,
     },
     './src/components/DashboardEditor/DashboardEditor.jsx': { branches: 65, functions: 71 },
     './src/components/BarChartCard/BarChartCard.jsx': {
@@ -46,41 +36,6 @@ module.exports = {
       // https://github.com/carbon-design-system/carbon-charts/issues/594
       functions: 77,
     },
-    './src/components/Table/TableHead/TableHeader.js': {
-      statements: 60,
-      branches: 50,
-      lines: 60,
-      functions: 33,
-    },
-    './src/components/Table/TableHead/ColumnResize.jsx': {
-      branches: 71,
-    },
-    './src/components/Table/TableHead/FilterHeaderRow/FilterHeaderRow.jsx': {
-      branches: 70,
-    },
-    './src/components/Table/TableBody/RowActionsCell/RowActionsCell.jsx': {
-      statements: 79,
-      branches: 70,
-      lines: 78,
-      functions: 75,
-    },
-    './src/components/Table/TableBody/RowActionsCell/RowActionsError.jsx': {
-      functions: 66,
-    },
-    './src/components/TileCatalog/CatalogContent.jsx': { branches: 50 },
-    './src/components/Dashboard/Dashboard.jsx': {
-      statements: 79,
-      branches: 50,
-      lines: 78,
-    },
-    './src/components/Dashboard/CardRenderer.jsx': {
-      statements: 51,
-      branches: 38,
-      lines: 51,
-      functions: 66,
-    },
-    './src/components/Page/PageHero.jsx': { branches: 77 },
-    './src/components/ColorDropdown/ColorDropdown.jsx': { branches: 75 },
   },
   globals: {
     __DEV__: false,