+++ packages/react/src/components/PageTitleBar/PageTitleBar.story.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
+import { select } from '@storybook/addon-knobs';
 import { Add24, TrashCan24 } from '@carbon/icons-react';
 import { spacing05 } from '@carbon/layout';
 import { Tabs, Tab } from 'carbon-components-react';
@@ -99,24 +100,26 @@ WithTooltipDescriptionWithNode.story = {
 };
 
 export const WithContent = () => (
-  <PageTitleBar
-    title={commonPageTitleBarProps.title}
-    description={commonPageTitleBarProps.description}
-    breadcrumb={pageTitleBarBreadcrumb}
-    content={
-      <Tabs>
-        <Tab label="Tab 1">
-          <div>Content for first tab.</div>
-        </Tab>
-        <Tab label="Tab 2">
-          <div>Content for second tab.</div>
-        </Tab>
-        <Tab label="Tab 3">
-          <div>Content for third tab.</div>
-        </Tab>
-      </Tabs>
-    }
-  />
+  <>
+    <PageTitleBar
+      title={commonPageTitleBarProps.title}
+      description={commonPageTitleBarProps.description}
+      breadcrumb={pageTitleBarBreadcrumb}
+      content={
+        <Tabs>
+          <Tab label="Tab 1">
+            <div style={{ height: '100rem' }}>Content for first tab.</div>
+          </Tab>
+          <Tab label="Tab 2">
+            <div>Content for second tab.</div>
+          </Tab>
+          <Tab label="Tab 3">
+            <div>Content for third tab.</div>
+          </Tab>
+        </Tabs>
+      }
+    />
+  </>
 );
 
 WithContent.story = {
@@ -206,6 +209,69 @@ export const WithEverything = () => (
 
 WithEverything.story = {
   name: 'with everything',
+};
+
+export const WithDynamicScrolling = () => (
+  <PageTitleBar
+    title={commonPageTitleBarProps.title}
+    description={commonPageTitleBarProps.description}
+    breadcrumb={pageTitleBarBreadcrumb}
+    headerMode={select('headerMode', ['DYNAMIC', 'STATIC', 'STICKY', 'CONDENSED'])}
+    headerModeDynamicOffSet={50} // need to offset the padding that storybook forces
+    extraContent={
+      <div>
+        <div
+          className="top"
+          style={{
+            display: 'flex',
+            flexDirection: 'row-reverse',
+          }}
+        />
+        <div style={{ display: 'flex' }} className="bottom">
+          <Button
+            renderIcon={Add24}
+            onClick={action('click')}
+            size="field"
+            hasIconOnly
+            iconDescription="Add"
+            kind="ghost"
+            tooltipPosition="bottom"
+            tooltipAlignment="center"
+          />
+          <Button
+            renderIcon={TrashCan24}
+            onClick={action('click')}
+            size="field"
+            hasIconOnly
+            iconDescription="Remove"
+            kind="ghost"
+            tooltipPosition="bottom"
+            tooltipAlignment="center"
+          />
+          <Button onClick={action('click')}>Take an action</Button>
+        </div>
+      </div>
+    }
+    editable
+    content={
+      <Tabs>
+        <Tab label="Tab 1">
+          <div style={{ height: '100rem' }}>Scroll me.</div>
+        </Tab>
+        <Tab label="Tab 2">
+          <div>Content for second tab.</div>
+        </Tab>
+        <Tab label="Tab 3">
+          <div>Content for third tab.</div>
+        </Tab>
+      </Tabs>
+    }
+    onEdit={action('edit')}
+  />
+);
+
+WithDynamicScrolling.story = {
+  name: 'With dynamic scrolling',
 };
 
 export const IsLoading = () => <PageTitleBar title={commonPageTitleBarProps.title} isLoading />;